+++ edison-togglz/src/main/java/de/otto/edison/togglz/controller/FeatureTogglesController.java
@@ -5,25 +5,27 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import static de.otto.edison.togglz.controller.TogglzRepresentation.togglzRepresentation;
+import static de.otto.edison.togglz.controller.FeatureTogglesRepresentation.togglzRepresentation;
 import static org.springframework.web.bind.annotation.RequestMethod.GET;
 
 @RestController
-public class TogglzInformationController {
+public class FeatureTogglesController {
 
     private final FeatureClassProvider featureClassProvider;
 
     @Autowired
-    public TogglzInformationController(final FeatureClassProvider featureClassProvider) {
+    public FeatureTogglesController(final FeatureClassProvider featureClassProvider) {
         this.featureClassProvider = featureClassProvider;
     }
 
     @RequestMapping(
-            value = "/internal/status/togglz",
-            produces = {"application/vnd.otto.monitoring.status+json", "application/json"},
+            value = "/internal/toggles",
+            produces = {
+                    "application/vnd.otto.monitoring.status+json",
+                    "application/json"},
             method = GET
     )
-    public TogglzRepresentation getStatusAsJson() {
+    public FeatureTogglesRepresentation getStatusAsJson() {
         return togglzRepresentation(featureClassProvider);
     }
 }